# Re-architect format.py in swift