--- a/SwiftSyntaxDevUtils/Sources/swift-syntax-dev-utils/SwiftSyntaxDevUtils.swift
+++ b/SwiftSyntaxDevUtils/Sources/swift-syntax-dev-utils/SwiftSyntaxDevUtils.swift
@@ -26,6 +26,7 @@ struct SwiftSyntaxDevUtils: ParsableCommand {
       """,
     subcommands: [
       Build.self,
+      Format.self,
       GenerateSourceCode.self,
       Test.self,
       VerifySourceCode.self,