Make a couple of changes to `format.py` and use the opportunity to re-write it in Swift.

1. Always build a local copy of swift-format in `swift-syntax/.swift-format-build` in release mode. This eliminates issues that occur if you had `swift-format` installed on your system and `format.py` would pick it up
2. Since the local swift-format build is persistent (doesn’t live in `/tmp`), we can build it in release. This increases the first run but decreases the time for any subsequent runs from ~15s to ~1s, which should pay usually pay off for frequent contributors because of faster turnaround times.
3. Add a little tip to add `swift-format` as a git hook to Contributing.md

And as an added bonus, the rewrite reduced the runtime of `format.(py|swift)` from 1.2s to 800ms.