@ahoppen 2023-08-07
https://github.com/apple/swift/pull/67777

@swift-ci Please test