@ahoppen 2023-08-09
https://github.com/apple/swift/pull/67777

@swift-ci Please test