@ahoppen 2023-08-18
https://github.com/apple/swift/pull/67777

@swift-ci Please test macOS