@ahoppen 2023-08-24
https://github.com/apple/swift/pull/67777

@swift-ci Please test macOS