@ahoppen 2023-08-25
https://github.com/apple/swift/pull/67777

@swift-ci Please test Windows