@ahoppen 2023-09-26
https://github.com/apple/swift/pull/67777

@swift-ci Please test macOS