@ahoppen 2023-09-27
https://github.com/apple/swift/pull/67777

@swift-ci Please test Linux