@ahoppen 2023-11-01
https://github.com/apple/swift/pull/67777

@swift-ci Please test macOS