@ahoppen 2023-11-01
apple/swift#67777

@swift-ci Please test macOS